refactor #1: Base, Domain 단위로 컴포넌트 역할 분리
