+++ src/components/index.js
@@ -1,4 +1,4 @@
 export { default as SkipNavigation } from './SkipNavigation/index.js';
-export { default as ScreenReader } from './ScreenReader/index.js';
-export { default as Typography } from './Typography/index.js';
-export { default as Footer } from './Footer/index.js';
+export { default as ScreenReader } from './Base/ScreenReader/index.js';
+export { default as Typography } from './Base/Typography/index.js';
+export { default as Footer } from './Domain/Footer/index.js';
